# dotnet-react/ClientApp/src/openapi/backendSchemas.ts
/**
 * Generated by @openapi-codegen
 *
 * @version 1.0
 */
export type Error = {
  /**
   * @minLength 1
   */
  code: string;
  /**
   * @minLength 1
   */
  message: string;
};

export type ErrorResponse = {
  error: Error;
};

export type WeatherForecast = {
  /**
   * @format date-time
   */
  date?: string;
  /**
   * @format int32
   */
  temperatureC?: number;
  /**
   * @format int32
   */
  temperatureF?: number;
  summary?: string | null;
};